Check the target directory actually exists before attempting to convert .po files

+++ scripts/localization/generate-php-files.sh
@@ -8,34 +8,44 @@ BUILDPATH=/tmp
 # Set the locales you want to convert
 LOCALES="ar at be bn cs de el es fr he hr it ka ko ku nl pl pt pt_BR ro ru sh sq sr sv uk zh zh_TW"
 
+cd ${BUILDPATH}/po/
+
 # Begin function that converts the files
 
 function convert2php {
   localization=$1
 
-cd ${BUILDPATH}/po/
+# Check that the target directory actually exists
 
-echo "Changing the ${localization} files to Newscoop style..."
+if [ -d ${localization} ]; then
 
-sed -i 's/# Newscoop translation file/<?php /g' ${localization}/*.po
-sed -i ':q;N;s/\n/ /g;t q' ${localization}/*.po
-sed -i 's/ msgid "/\nregGS("/g' ${localization}/*.po
-sed -i 's/ msgstr "/, "/g' ${localization}/*.po
-sed -i '/<?php/!s/[ ]*$/);/g' ${localization}/*.po
-sed -i '$ a\?>' ${localization}/*.po
+  echo "Changing the ${localization} files to Newscoop style..."
 
-# Strip the final newline from the file
+  sed -i 's/# Newscoop translation file/<?php /g' ${localization}/*.po
+  sed -i ':q;N;s/\n/ /g;t q' ${localization}/*.po
+  sed -i 's/ msgid "/\nregGS("/g' ${localization}/*.po
+  sed -i 's/ msgstr "/, "/g' ${localization}/*.po
+  sed -i '/<?php/!s/[ ]*$/);/g' ${localization}/*.po
+  sed -i '$ a\?>' ${localization}/*.po
 
-for i in ${localization}/*.po; do
- awk '{q=p;p=$0}NR>1{print q}END{ORS = ""; print p}' $i > $i.tmp
- rm $i
-done
+  # Strip the final newline from the file
 
-# Give the changed files a .php extension
+  for i in ${localization}/*.po; do
+   awk '{q=p;p=$0}NR>1{print q}END{ORS = ""; print p}' $i > $i.tmp
+   rm $i
+  done
 
-for i in ${localization}/*.po.tmp; do
- mv "$i" "${i/.po.tmp}".php
-done
+  # Give the changed files a .php extension
+
+  for i in ${localization}/*.po.tmp; do
+   mv "$i" "${i/.po.tmp}".php
+  done
+
+else
+
+  echo "The ${localization} directory was not found."
+
+fi
 
 }
